move c# object models into EmbeddedExamples.cs file rather than separate files
