--- examples/dotnet/Examples/Address.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using MongoDB.Bson;
-using Realms;
-using Realms.Sync;
-
-namespace Examples
-{
-    public class Address : EmbeddedObject
-    {
-        public string Street { get; set; }
-        public string City { get; set; }
-        public string Country { get; set; }
-        public string PostalCode { get; set; }
-
-        public Address() { }
-    }
-}
