--- examples/dotnet/Examples/Business.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using MongoDB.Bson;
-using Realms;
-using Realms.Sync;
-
-namespace Examples
-{
-    public class Business : RealmObject
-    {
-        [PrimaryKey]
-        [MapTo("_id")]
-        public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-
-        public string Name { get; set; }
-        public IList<Address> addresses { get; }
-    }
-}
-
