+++ examples/dotnet/Examples/EmbeddedExamples.cs
@@ -141,5 +141,40 @@ namespace Examples
             }
             return;
         }
+
+        // :code-block-start:embedded-classes
+        public class Address : EmbeddedObject
+        {
+            public string Street { get; set; }
+            public string City { get; set; }
+            public string Country { get; set; }
+            public string PostalCode { get; set; }
+
+            public Address() { }
+        }
+        public class Contact : RealmObject
+        {
+
+            [PrimaryKey]
+            [MapTo("_id")]
+            public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
+
+            public string Name { get; set; }
+
+            public Address Address { get; set; } // embed a single address 
+
+            public Contact() { }
+
+        }
+        public class Business : RealmObject
+        {
+            [PrimaryKey]
+            [MapTo("_id")]
+            public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
+
+            public string Name { get; set; }
+            public IList<Address> addresses { get; }
+        }
+        //:code-block-end:
     }
 }
