+++ source/index.txt
@@ -227,6 +227,13 @@ documentation for your platform, framework, and language of choice:
 
    Web SDK </web>
 
+.. toctree::
+   :titlesonly:
+   :caption: Dot New SDK
+   :hidden:
+
+   Dot Net SDK </dotnet>
+
 Billing & Reference
 -------------------
